requirements.txt
torch==2.8.0
tensorboard==2.20.0
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
tqdm==4.67.1
